+++ modules/core/src/directive/dynamic-template.directive.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed, ComponentFixture } from "@angular/core/testing";
 import { Component, Type, DebugElement } from "@angular/core";
 import { By } from "@angular/platform-browser";
-import { DynamicTemplateDirective } from "./dynamic-template.directive";
+import { DynamicTemplateDirective, DYNAMIC_TEMPLATE_DIRECTIVE_ALIGN_END } from "./dynamic-template.directive";
 
 @Component({
     template: `
@@ -14,10 +14,13 @@ class TestComponent {
 
 describe("DynamicTemplateDirective test suite", () => {
 
-    let fixture: ComponentFixture<TestComponent>,
+    let directive: DynamicTemplateDirective,
+        fixture: ComponentFixture<TestComponent>,
         directives: DebugElement[];
 
     beforeEach(() => {
+
+        directive = new DynamicTemplateDirective(null);
 
         fixture = TestBed.configureTestingModule({
 
@@ -28,5 +31,13 @@ describe("DynamicTemplateDirective test suite", () => {
         fixture.detectChanges();
 
         directives = fixture.debugElement.queryAll(By.directive(DynamicTemplateDirective));
+    });
+
+    it("should be initialized correctly", () => {
+
+        expect(directive.align).toEqual(DYNAMIC_TEMPLATE_DIRECTIVE_ALIGN_END);
+        expect(directive.modelId).toBeUndefined();
+        expect(directive.modelType).toBeUndefined();
+        expect(directive.type).toBeNull();
     });
 });