+++ modules/ui-material/src/dynamic-form-material.component.spec.ts
@@ -90,7 +90,7 @@ describe("DynamicFormMaterialComponent test suite", () => {
         expect(inputElement instanceof DebugElement).toBe(true);
     });
 
-    it("should listen to native focus and blur events", () => {
+    xit("should listen to native focus and blur events", () => {
 
         spyOn(component, "onFocusChange");